+++ src/ResponseBuilderBase.php
@@ -58,9 +58,16 @@ abstract class ResponseBuilderBase
 	public const CONF_KEY_MAP                    = self::CONF_CONFIG . '.map';
 	public const CONF_KEY_ENCODING_OPTIONS       = self::CONF_CONFIG . '.encoding_options';
 	public const CONF_KEY_CONVERTER              = self::CONF_CONFIG . '.converter';
+	public const CONF_KEY_CONVERTER_CLASSES      = self::CONF_KEY_CONVERTER . '.classes';
+	public const CONF_KEY_CONVERTER_PRIMITIVES   = self::CONF_KEY_CONVERTER . '.primitives';
 	public const CONF_KEY_MIN_CODE               = self::CONF_CONFIG . '.min_code';
 	public const CONF_KEY_MAX_CODE               = self::CONF_CONFIG . '.max_code';
 	public const CONF_KEY_EXCEPTION_HANDLER      = self::CONF_CONFIG . '.exception_handler';
+
+	public const PRIMITIVE_ARRAY   = 'array';
+	public const PRIMITIVE_BOOLEAN = 'boolean';
+	public const PRIMITIVE_DOUBLE  = 'double';
+	public const PRIMITIVE_INT     = 'int';
 
 	/**
 	 * Default keys to be used by exception handler while adding debug information