+++ tests/tests/ResponseBuilder/SuccessTest.php
@@ -13,6 +13,7 @@ namespace MarcinOrlowski\ResponseBuilder\Tests;
  * @link      https://github.com/MarcinOrlowski/laravel-api-response-builder
  */
 
+use Illuminate\Support\Facades\Config;
 use MarcinOrlowski\ResponseBuilder\BaseApiCodes;
 use MarcinOrlowski\ResponseBuilder\ResponseBuilder;
 
@@ -53,7 +54,14 @@ class SuccessTest extends TestCase
 		$j = $this->getResponseSuccessObject(BaseApiCodes::OK());
 
 		$this->assertNotNull($j->data);
-		$this->assertArraysEquals($payload, (array)$j->data);
+		$data = (array)$j->data;
+
+		$cfg = Config::get(ResponseBuilder::CONF_KEY_CONVERTER_PRIMITIVES);
+		$this->assertNotEmpty($cfg);
+		$key = $cfg[ ResponseBuilder::PRIMITIVE_ARRAY ][ ResponseBuilder::KEY_KEY ];
+
+		$this->assertCount(1, $data);
+		$this->assertArraysEquals($payload, (array)$j->data->{$key});
 		$this->assertEquals(\Lang::get(BaseApiCodes::getCodeMessageKey(BaseApiCodes::OK())), $j->message);
 	}