+++ CHANGES.md
@@ -6,6 +6,12 @@ See [compatibility docs](docs/compatibility.md) for details about backward compa
 
 ## CHANGE LOG ##
 
+* @dev
+   * **BACKWARD INCOMPATIBLE CHANGES** ([more info](docs/compatibility.md))
+   * [RB-156] Added logic to deal with directly returned objects or arrays.
+   * Passing primitives as direct payload (i.e. `success(12.50);` is now supported for `array`, `boolean`,
+     `double`, `integer` and `string` types, configurable via new `converter/primitives`.
+
 * v8.1.1 (2020-10-15)
    * [RB-155] Fixed `ResponseBuilder` internals preventing exdending class code from
      being invoked, thus making response object structure manipulation ineffective (reported by krek95)